feat: benchmark on multiple test sets

+++ raffle_ds_research/core/workflows/schedule.py
@@ -4,11 +4,11 @@ import pydantic
 
 
 class Schedule(Protocol):
-    def __call__(self, step: int, start: float, end: float, period: int, offset: int = 0) -> float:
+    def __call__(self, step: float, start: float, end: float, period: int, offset: int = 0) -> float:
         ...
 
 
-def LinearSchedule(step: int, start: float, end: float, period: int, offset: int = 0) -> float:
+def LinearSchedule(step: float, start: float, end: float, period: int, offset: int = 0) -> float:
     if step < offset:
         return start
     elif step >= offset + period:
@@ -32,6 +32,6 @@ class ScheduleConfig(pydantic.BaseModel):
     period: int
     offset: int = 0
 
-    def __call__(self, step: int) -> float:
+    def __call__(self, step: float) -> float:
         schedule: Schedule = SCHEDULES[self.type]
         return schedule(step, self.start, self.end, self.period, self.offset)
